+++ anymail/webhooks/mailtrap.py
@@ -1,6 +1,15 @@
+import hashlib
+import hmac
 import json
 import sys
 from datetime import datetime, timezone
+
+from django.utils.crypto import constant_time_compare
+
+from ..exceptions import AnymailWebhookValidationFailure
+from ..signals import AnymailTrackingEvent, EventType, RejectReason, tracking
+from ..utils import get_anymail_setting
+from .base import AnymailBaseWebhookView
 
 if sys.version_info < (3, 11):
     from typing import Literal
@@ -9,12 +18,50 @@ if sys.version_info < (3, 11):
 else:
     from typing import Literal, NotRequired, TypedDict
 
-from ..signals import AnymailTrackingEvent, EventType, RejectReason, tracking
-from .base import AnymailBaseWebhookView
+
+class MailtrapWebhookView(AnymailBaseWebhookView):
+    esp_name = "Mailtrap"
+
+    def __init__(self, _secret_name, **kwargs):
+        signing_secret = get_anymail_setting(
+            _secret_name,
+            esp_name=self.esp_name,
+            default=None,
+            kwargs=kwargs,
+        )
+        if signing_secret is None:
+            self.signing_secret = None
+        else:
+            self.signing_secret = signing_secret.encode()
+            self.warn_if_no_basic_auth = False
+        self._secret_setting_name = f"{self.esp_name}_{_secret_name}".upper()
+        super().__init__(**kwargs)
+
+    def validate_request(self, request):
+        if self.signing_secret is None:
+            return
+
+        try:
+            signature = request.headers["Mailtrap-Signature"]
+        except KeyError:
+            raise AnymailWebhookValidationFailure(
+                "Mailtrap webhook called without signature"
+            ) from None
+
+        expected_signature = hmac.new(
+            key=self.signing_secret,
+            msg=request.body,
+            digestmod=hashlib.sha256,
+        ).hexdigest()
+        if not constant_time_compare(signature, expected_signature):
+            raise AnymailWebhookValidationFailure(
+                f"Mailtrap webhook called with incorrect signature"
+                f" (check Anymail {self._secret_setting_name} setting)"
+            )
 
 
-class MailtrapEvent(TypedDict):
-    # https://api-docs.mailtrap.io/docs/mailtrap-api-docs/016fe2a1efd5a-receive-events-json-format
+class MailtrapReceiveEvent(TypedDict):
+    # https://docs.mailtrap.io/email-api-smtp/advanced/webhooks
     event: Literal[
         "delivery",
         "open",
@@ -42,16 +89,21 @@ class MailtrapEvent(TypedDict):
     url: NotRequired[str]
 
 
-class MailtrapTrackingWebhookView(AnymailBaseWebhookView):
+class MailtrapTrackingWebhookView(MailtrapWebhookView):
     """Handler for Mailtrap delivery and engagement tracking webhooks"""
 
-    esp_name = "Mailtrap"
     signal = tracking
 
+    # (Declaring class attr allows override by kwargs in View.as_view.)
+    tracking_secret = None
+
+    def __init__(self, **kwargs):
+        super().__init__(_secret_name="tracking_secret", **kwargs)
+
     def parse_events(self, request):
-        esp_events: list[MailtrapEvent] = json.loads(request.body.decode("utf-8")).get(
-            "events", []
-        )
+        esp_events: list[MailtrapReceiveEvent] = json.loads(
+            request.body.decode("utf-8")
+        ).get("events", [])
         return [self.esp_to_anymail_event(esp_event) for esp_event in esp_events]
 
     # https://help.mailtrap.io/article/87-statuses-and-events
@@ -79,7 +131,7 @@ class MailtrapTrackingWebhookView(AnymailBaseWebhookView):
         "soft bounce": RejectReason.OTHER,
     }
 
-    def esp_to_anymail_event(self, esp_event: MailtrapEvent):
+    def esp_to_anymail_event(self, esp_event: MailtrapReceiveEvent):
         event_type = self.event_types.get(esp_event["event"], EventType.UNKNOWN)
         timestamp = datetime.fromtimestamp(esp_event["timestamp"], tz=timezone.utc)
         reject_reason = self.reject_reasons.get(esp_event["event"])